fix(comparability): validate both hybrid legs; split language-mode carve-outs into a sibling module
Codex review on PR #816 (P2, eighth round on the same carve-out):
fresh evidence beyond the prior hybrid-corroboration fix -- a real
hybrid snapshot carries BOTH castxml_* and clang_* provenance, but the
previous fix's _tc() fallback only ever consulted the castxml leg. If
castxml's own leg stays unchanged while the clang frontend/compiler
changes between two snapshots, the castxml-only checks passed and the
carve-out discarded the combined compiler_version mismatch -- the
only profile field carrying the changed clang identity -- reporting
the pair comparable despite using different extraction toolchains on
that leg.

Fixed with a new _tc_match() helper: for an ordinary (bare-keyed)
snapshot it's the same non-empty-and-equal check as before; for a
hybrid-merged snapshot (no bare key at all) it requires EVERY
castxml_/clang_ leg present on either side to independently
corroborate. Applied to producer, frontend version, and frontend
sha256 -- the three fields this carve-out already required to match
unconditionally, and exactly the fields a clang-leg identity drift
would show up in. The compiler_version/target_triple/install-dir
checks specific to castxml's own gcc/g++ host-compiler-split trick
stay castxml-leg-only via the existing _tc() fallback: that heuristic
has no clang equivalent (clang is both frontend and "compiler" for
its own leg, and its banner never splits by language mode the way
castxml's resolved host compiler does).

This finding needed real new code in a file that was already sitting
at the exact 2000-line AI-readiness hard cap with no docstring
trimming left to absorb it -- the point flagged in this session's own
prior commits as the trigger to stop squeezing and split instead.
Split _language_standard_probe_upgrade_corroborated,
_language_standard_content_divergence_corroborated, and their shared
helpers/constants (_tc, _tc_match, _compiler_install_dir,
_compiler_version_sans_driver_name, _driver_prefix,
_is_gcc_gxx_driver_pair, _nonempty_match,
_language_standard_is_lang_pinned,
_language_standard_is_known_auto_resolved_form,
_PROBED_STANDARD_PREFIX, _FORCED_C_STANDARD, _FORCED_CPP20_STANDARD,
_KNOWN_AUTO_RESOLVED_BARE_STANDARDS, _UNRESOLVED_STANDARD_SPELLINGS)
into a new sibling module, abicheck/comparability_language_mode.py --
following this same file's own established pattern
(comparability_fields.py/comparability_json.py/
comparability_sequences.py already split off other concerns the
identical way). The two corroboration functions lost their leading
underscore (now the module's public API); comparability.py imports
them plus re-exports of _is_gcc_gxx_driver_pair/_PROBED_STANDARD_PREFIX
for import-path stability, since
tests/test_comparability_gate_probe_upgrade.py already imports both
directly from abicheck.comparability and predates this split.
comparability.py is now 1524 lines (down from 2000); the new module is
551 lines -- both comfortably under the cap, with real headroom
restored for future findings instead of sitting at the wire.

Regression coverage:
test_gate_content_driven_still_raises_for_hybrid_clang_leg_drift
constructs exactly Codex's counterexample -- a hybrid pair whose
castxml legs are identical but whose clang legs differ in
version/sha256 -- and confirms it still raises. All 179 existing
comparability/hybrid/ast-provenance tests pass unchanged after the
split; ruff/mypy clean on both files; no import cycle
(check_ai_readiness.py's import-cycle-growth check passes).
